C++ Hospital Management Project
Made a project on C++. on managing the details of the patients who have been admitted. (The doctor who is treating the patient and the tests conducted on the patient is stored in a file which helps in generating the bill at the time of discharge.)

The program stores the details of the tests, doctors and rooms available in the hospital. We can view the room details, add new rooms, remove or modify details of a particular room.

It also accepts all the details of the doctors which can be modified or deleted.

All in all the application provides robust and friendly environment for the users to navigate through a Hospital Room Management.